# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/profile": {
      "get": {
        "description": "현재 로그인한 사용자의 정보를 조회합니다.",
        "operationId": "UsersController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "사용자 정보 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "현재 사용자 정보 조회",
        "tags": [
          "users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "description": "특정 ID를 가진 사용자 정보를 조회합니다.",
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "사용자 정보 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "summary": "ID로 사용자 조회",
        "tags": [
          "users"
        ]
      }
    },
...
  },
  "info": {
    "title": "COCOMOMO API",
    "description": "COCOMOMO 백엔드 API 문서",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "사용자 ID (자동생성)",
            "example": 1
          },
          "githubId": {
            "type": "number",
            "description": "GitHub ID",
            "example": 12345678
          },
          "username": {
            "type": "string",
            "description": "사용자 이름",
            "example": "username"
          },
          "email": {
            "type": "string",
            "description": "사용자 이메일",
            "example": "user@example.com"
          },
          "avatarUrl": {
            "type": "string",
            "description": "프로필 이미지 URL",
            "example": "https://avatars.githubusercontent.com/u/1234567"
          },
          "role": {
            "type": "string",
            "description": "사용자 권한",
            "enum": [
              "user",
              "admin"
            ],
            "example": "user"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "사용자 계정 생성 시간",
            "example": "2025-06-09T13:30:51+09:00"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "사용자 정보 수정 시간",
            "example": "2025-06-09T13:30:51+09:00"
          }
        },
        "required": [
          "id",
          "githubId",
          "username",
          "role",
          "createdAt",
          "updatedAt"
        ]
      },
...
    }
  }
}